KumsalTR/core/calls.py
# @The_Team_kumsal tarafından yasal olarak geliştirildi keyifli kullanımlar #kumsalteam
# Copyright (c) 2025 TheHamkerAlone
# Licensed under the MIT License.
# This file is part of KumsalTR

import asyncio
from ntgcalls import (ConnectionNotFound, TelegramServerError,
                      RTMPStreamingUnsupported)
from pyrogram.errors import MessageIdInvalid
from pyrogram.types import Message
from pytgcalls import PyTgCalls, exceptions, types
from pytgcalls.pytgcalls_session import PyTgCallsSession

from KumsalTR import app, config, db, lang, logger, queue, userbot, yt
from KumsalTR.helpers import Media, Track, buttons


class TgCall(PyTgCalls):
    def __init__(self):
        self.clients = []

    async def pause(self, chat_id: int) -> bool:
        client = await db.get_assistant(chat_id)
        if not client:
            return False
        await db.playing(chat_id, paused=True)
        return await client.pause(chat_id)

    async def resume(self, chat_id: int) -> bool:
        client = await db.get_assistant(chat_id)
        if not client:
            return False
        await db.playing(chat_id, paused=False)
        return await client.resume(chat_id)

    async def stop(self, chat_id: int) -> None:
        client = await db.get_assistant(chat_id)
        try:
            queue.clear(chat_id)
            await db.remove_call(chat_id)
        except:
            pass

        if client:
            try:
                await client.leave_call(chat_id, close=False)
            except:
                pass

    async def play_media(
        self,
        chat_id: int,
        message: Message,
        media: Media | Track,
        seek_time: int = 0,
    ) -> None:
        client = await db.get_assistant(chat_id)
        _lang = await lang.get_lang(chat_id)

        if not client:
            return await message.edit_text("<b>❌ Hᴀᴛᴀ:</b> Asɪsᴛᴀɴʟᴀʀ ʜᴇɴᴜ̈ᴢ ʜᴀᴢɪʀ ᴅᴇɢɪʟ. Lᴜ̈ᴛғᴇɴ ʙɪʀᴀᴢ ʙᴇᴋʟᴇʏɪɴ.")

        if not media.file_path:
            await message.edit_text(_lang["error_no_file"].format(config.SUPPORT_CHAT))
            return await self.play_next(chat_id)

        stream = types.MediaStream(
            media_path=media.file_path,
            audio_parameters=types.AudioQuality.HIGH,
            video_parameters=types.VideoQuality.HD_720p,
            audio_flags=types.MediaStream.Flags.REQUIRED,
            video_flags=(
                types.MediaStream.Flags.AUTO_DETECT
                if media.video
                else types.MediaStream.Flags.IGNORE
            ),
            ffmpeg_parameters=f"-ss {seek_time}" if seek_time > 1 else None,
        )
        try:
            try:
                await asyncio.wait_for(
                    client.play(
                        chat_id=chat_id,
                        stream=stream,
                        config=types.GroupCallConfig(auto_start=False),
                    ),
                    timeout=20
                )
            except asyncio.TimeoutError:
                await message.edit_text("<b>❌ Hᴀᴛᴀ:</b> Sᴛʀᴇᴀᴍ ʙᴀşʟᴀᴛɪʟɪʀᴋᴇɴ ᴢᴀᴍᴀɴ ᴀşɪᴍɪ ᴏʟᴜşᴛᴜ. Lᴜ̈ᴛғᴇɴ ᴛᴇᴋʀᴀʀ ᴅᴇɴᴇʏɪɴ.")
                return await self.stop(chat_id)

            if not seek_time:
                media.time = 1
                await db.add_call(chat_id)
                text = _lang["play_media"].format(
                    media.url,
                    media.title,
                    media.duration,
                    media.user,
                )
                keyboard = buttons.controls(chat_id)
                try:
                    # Update Stats
                    if media.user_id:
                        await db.update_stats(media.user_id, media.title)
                        # Notify soulmate
                        partner_id = await db.get_soulmate(media.user_id)
                        if partner_id:
                            try:
                                await app.send_message(
                                    partner_id,
                                    f"<b><tg-emoji emoji-id=\"5470135030393090150\">💖</emoji> 𝐑𝐮𝐡 𝐄𝐬̧𝐢𝐧 𝐒̧𝐮 𝐀𝐧 𝐌𝐮̈𝐳𝐢𝐤 𝐃𝐢𝐧𝐥𝐢𝐲𝐨𝐫!</b>\n\n• <b>Şarkı:</b> {media.title}\n• <b>Grup:</b> {message.chat.title or 'Özel'}"
                                )
                            except: pass

                    # Fotoğraf yerine düz metin olarak düzenle
                    await message.edit_text(
                        text=text,
                        reply_markup=keyboard,
                        disable_web_page_preview=True,
                    )
                except MessageIdInvalid:
                    # Mesaj silinmişse fotoğrafsız yeni mesaj gönder
                    media.message_id = (await app.send_message(
                        chat_id=chat_id,
                        text=text,
                        reply_markup=keyboard,
                        disable_web_page_preview=True,
                    )).id
        except FileNotFoundError:
            await message.edit_text(_lang["error_no_file"].format(config.SUPPORT_CHAT))
            await self.play_next(chat_id)
        except exceptions.NoActiveGroupCall:
            await self.stop(chat_id)
            await message.edit_text(_lang["error_no_call"])
        except exceptions.NoAudioSourceFound:
            await message.edit_text(_lang["error_no_audio"])
            await self.play_next(chat_id)
        except (ConnectionNotFound, TelegramServerError):
            await self.stop(chat_id)
            await message.edit_text(_lang["error_tg_server"])
        except RTMPStreamingUnsupported:
            await self.stop(chat_id)
            await message.edit_text(_lang["error_rtmp"])

    async def replay(self, chat_id: int) -> None:
        if not await db.get_call(chat_id):
            return

        media = queue.get_current(chat_id)
        _lang = await lang.get_lang(chat_id)
        msg = await app.send_message(chat_id=chat_id, text=_lang["play_again"])
        await self.play_media(chat_id, msg, media)

    async def play_next(self, chat_id: int) -> None:
        media = queue.get_next(chat_id)

        if not media:
            return await self.stop(chat_id)

        try:
            if media.message_id:
                await app.delete_messages(
                    chat_id=chat_id,
                    message_ids=media.message_id,
                    revoke=True,
                )
                media.message_id = 0
        except:
            pass

        _lang = await lang.get_lang(chat_id)
        msg = await app.send_message(chat_id=chat_id, text=_lang["play_next"])
        if not media.file_path:
            media.file_path = await yt.download(media.id, video=media.video)
            if not media.file_path:
                await self.stop(chat_id)
                return await msg.edit_text(
                    _lang["error_no_file"].format(config.SUPPORT_CHAT)
                )

        media.message_id = msg.id
        await self.play_media(chat_id, msg, media)

    async def ping(self) -> float:
        pings = [client.ping for client in self.clients]
        return round(sum(pings) / len(pings), 2)

    async def decorators(self, client: PyTgCalls) -> None:
        for client in self.clients:
            @client.on_update()
            async def update_handler(_, update: types.Update) -> None:
                if isinstance(update, types.StreamEnded):
                    if update.stream_type == types.StreamEnded.Type.AUDIO:
                        await self.play_next(update.chat_id)
                elif isinstance(update, types.ChatUpdate):
                    if update.status in [
                        types.ChatUpdate.Status.KICKED,
                        types.ChatUpdate.Status.LEFT_GROUP,
                        types.ChatUpdate.Status.CLOSED_VOICE_CHAT,
                    ]:
                        await self.stop(update.chat_id)

    async def boot(self) -> None:
        PyTgCallsSession.notice_displayed = True
        for ub in userbot.clients:
            client = PyTgCalls(ub, cache_duration=100)
            await client.start()
            self.clients.append(client)
            await self.decorators(client)
        logger.info("PyTgCalls client(s) started without thumbnails.")

